Added error surfacing for setupForDecode(). Removed warnings on Windows. Reworked functionality to play nice on Windows. Added note to CTimer about callback thread. Minor style updates. Added new typedef for T2DSize. Significant rework in CMediaFoundationServices on Windows. Implemented CMediaFoundationResampler on Windows (renamed from CMediaFoundationAudioConverter). Reworked CMediaFoundationAudioCodecs and CH264VideoCodec on Windows based on other changes. Significant fixes and improvements for CAudioPlayer-Windows.

+++ Source/Media/Media Sources/CQuickTimeMediaSource.cpp
@@ -271,7 +271,7 @@ struct SQTAudioSampleDescription {
 								switch (EndianU16_BtoN(mVersion)) {
 									case 0:		return EndianU16_BtoN(_.mV0.mChannels);
 									case 1:		return EndianU16_BtoN(_.mV1.mChannels);
-									case 2:		return EndianU32_BtoN(_.mV2.mChannels);
+									case 2:		return (UInt16) EndianU32_BtoN(_.mV2.mChannels);
 									default:	return 0;
 								}
 							}
